refactor: resolve medium findings and extract workflow-editor sub-components
7. saveSnapshot stable identity — moved editedSteps/Transitions into refs
   (synced by useEffect) so saveSnapshot reads current state without
   capturing them in its dep array, eliminating cascading re-creations of
   addStep, removeStep, and moveStep.

8. definitions.ts: removed double cast `as Omit<WorkflowDefinition,...>` —
   saveWorkflowDefinition already validates through Zod internally, so the
   call-site now passes `unknown` and the function parameter is typed
   `unknown`. The intermediate misleading assertion is gone.

9. YAML Apply now validates both steps (WorkflowStepSchema.array()) and
   transitions (TransitionSchema.array()) through Zod instead of only
   checking that steps is an array. Errors surface the first Zod issue
   message so the user knows exactly what to fix.

10. (already done) console.warn on legacy parse failure.

11. Extracted StepEditor, CoworkSection, StepIdField into a
    workflow-editor/ sub-directory:
    - constants.ts    — AUTONOMY_LEVELS, STEP_TYPES, KNOWN_MODELS, etc.
    - step-editor-fields.tsx  — EditableField, Section primitives
    - cowork-section.tsx      — CoworkSection + CoworkOutputSchemaEditor
    - step-editor.tsx         — StepEditor, StepIdField, buildExecutorChangePatch, toSlug

    workflow-editor-canvas.tsx shrinks from 1437 → 568 lines.
