+++ packages/platform-ui/src/components/workflows/workflow-editor/step-editor-fields.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+import React from 'react';
+import { cn } from '@/lib/utils';
+
+export function EditableField({ label, value, onChange, mono, placeholder, suffix }: {
+  label: string; value: string; onChange: (v: string) => void;
+  mono?: boolean; placeholder?: string; suffix?: string;
+}) {
+  return (
+    <div className="flex items-baseline justify-between gap-3">
+      <span className="text-xs text-muted-foreground shrink-0">{label}</span>
+      <div className="flex items-baseline gap-1">
+        <input
+          value={value}
+          onChange={(e) => onChange(e.target.value)}
+          placeholder={placeholder}
+          className={cn(
+            'bg-transparent text-xs text-right border-0 border-b border-transparent hover:border-muted-foreground/20 focus:border-primary px-0 py-0 focus:outline-none transition-colors w-28',
+            mono && 'font-mono',
+            !value && 'placeholder:text-muted-foreground/40 placeholder:italic',
+          )}
+        />
+        {suffix && <span className="text-[10px] text-muted-foreground">{suffix}</span>}
+      </div>
+    </div>
+  );
+}
+
+export function Section({ title, children }: { title: string; children: React.ReactNode }) {
+  return (
+    <div className="space-y-1.5">
+      <p className="text-[11px] font-semibold uppercase tracking-widest text-muted-foreground/60">{title}</p>
+      {children}
+    </div>
+  );
+}
